requirements.txt
numpy<2
streamlit
rembg
onnxruntime
pillow
opencv-python
scikit-learn